+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.14.5
 
 * `Closeable` is now in __all__
+* `Closeable` has __slots__